Add files via upload
Background 
XYZ is a private equity firm in US. Due to remarkable growth in the Cab Industry in last few years and multiple key players in the market, it is planning for an investment in Cab industry.
• Objective: Provide actionable insights to help XYZ firm in identifying the right company for making investment.
 The analysis has been divided into four parts:
• Data Understanding
• Forecasting profit and number of rides for each cab type
• Finding the most profitable Cab company
• Recommendations for investment
Data exploration 
All four data sources have been combined into “All data.”
Assumptions 
Profit is calculated as the difference between the price charged and the cost of the trip.
The "users" feature is the number of customers willing to use either of the cabs.
Any Nulls are treated as outliers.
Recommendations 
The following are the outcomes for both cabs
Profit: Yellow Cab is doing better regarding profits and customer reach. 
Income reach: Cabs are preferred by the rich and middle class, with Yellow Cabs performing better in all income groups.
Profit forecasting – Yellow Cab is projected to hit $13,581,226 in profit by 2020. 
Distance- most customers prefer rides that are less than 35 kilometers. 
Day-wise, Sundays and Mondays have more requests than other days. 
Cities such as Washington, DC, and Boston, MA, have the highest number of users, with many customers preferring Yellow Cab. 

 Based on the analysis, Yellow Cab is a better investment option.